{"schema":"decodex/commit/1","summary":"Split external scoreboard modules","authority":"manual"}
